Updated The structure
So now there are 2 channels per Node, one for receiving and one for sending.
I also created another Node data structure, which I guess is functionally the same as this one, its just on another file.
Tseng was saying something about using a network to do the sending and receiving, but this is a little simpler. Uhh, still need to test it out and change a couple things, but I think actually pushing should work now.